W I P
Before, `./test.sh && pidof $(which python)` will give a list of pids
which means we are not properly killing child processes. It is probably
a result of the warning RE pending tasks.

Now with this this WIP, we do not see zombie child processes anymore.
We still see the "Task was destroyed but it is pending" warning. We
also see a new error "asyncio.exceptions.CancelledError". This is
considered progress.

We now need to handle this new error and ensure we do not see that
warning anymore.