(#26) Fix ./publish.sh not publish on origin/main branch
Previously on origin/main or main branch, we're getting
old and new version by running `git diff origin/main`. This
works for non-main branch but doesn't work for origin/main
or main branch. Instead, we need to use `git diff HEAD~1`
which compares the version of the current commit with the previous
one. This is okay since in this repository we only allow
squashed commit when merging from feature branch.

We also need to make sure we log correctly:
* Remove the ~12 hours waiting log when on origin/main or main
  branch since that does not apply.
* Rename repository `aiotaskq` and `aiotaskqdev` to `pypi` and
  `pypitest` to remove ambiguity and improve clarity in the
  logs.