(#26) Fix not auto-publishing on origin/main
Previously we get the name of the current branch
by using `git branch ...` but it did not give
an easily parse-able result, and we got the parsing
wrong. Because we got the wrong branch name,
we got the wrong version diff, and hence not trigger
publishing to pypi.

Now, we use a more cleaner command: `git name-rev
--name-only HEAD` and we should be getting the
correct version diff and hence trigger publishing
to pypi correctly.